Refactoring 6. Some small fixes
