+++ app/src/main/java/com/example/playlistmaker/domain/usecases/GetSearchResultTrackListUseCase.kt
@@ -3,7 +3,7 @@ package com.example.playlistmaker.domain.usecases
 import com.example.playlistmaker.domain.model.Track
 import com.example.playlistmaker.domain.repository.SearchRepository
 
-class GetAlreadySearchTracksUseCase(private val searchRepository: SearchRepository) {
+class GetSearchResultTrackListUseCase(private val searchRepository: SearchRepository) {
     fun execute(): List<Track> {
         return searchRepository.getSearchTracks()
     }
