+++ app/src/main/java/com/example/playlistmaker/domain/usecases/MediaPlayerInitUseCase.kt
@@ -1,10 +1,26 @@
 package com.example.playlistmaker.domain.usecases
 
-interface MediaPlayerInitUseCase {
-    fun initPlayer(urlForMusicPreview: String)
-    fun setListeners(
+import com.example.playlistmaker.domain.libraries.MediaPlayerContract
+import com.example.playlistmaker.domain.model.PlayerStatus
+
+class MediaPlayerInitUseCase(
+    private val mediaPlayer: MediaPlayerContract,
+    private val callback: (PlayerStatus) -> (Unit)
+) {
+    fun initPlayer(urlForMusicPreview: String) {
+        setListeners({ callback.invoke(PlayerStatus.STATE_PREPARED) },
+            { callback.invoke(PlayerStatus.STATE_PREPARED) },
+            { callback.invoke(PlayerStatus.STATE_ERROR) })
+        mediaPlayer.initMediaPlayer(urlForMusicPreview)
+    }
+
+    private fun setListeners(
         onPreparedListener: () -> Unit,
         onCompletionListener: () -> Unit,
         onOnErrorListener: () -> Unit,
-    )
+    ) {
+        mediaPlayer.setOnPreparedListener(onPreparedListener)
+        mediaPlayer.setOnCompletionListener(onCompletionListener)
+        mediaPlayer.setOnErrorListener(onOnErrorListener)
+    }
 }
